eval update

+++ rely/evaluate/pum_eval.py
@@ -20,7 +20,7 @@ def load_model_and_tokenizer(checkpoint_path: str):
     
     # Load tokenizer from checkpoint
     tokenizer = AutoTokenizer.from_pretrained(checkpoint_path)
-    model = SoftClassificationPRMModel.from_pretrained(checkpoint_path, dtype=torch.float32)
+    model = SoftClassificationPRMModel.from_pretrained(checkpoint_path, dtype=torch.bfloat16)
     
     return model, tokenizer
 
@@ -64,7 +64,7 @@ def evaluate_model(model, tokenizer, dataset, device, max_length, step_separator
 
                 # Convert to tensors and move to device
                 input_ids = torch.tensor([tokenized["input_ids"]], device=device)
-                labels_tensor = torch.tensor([tokenized["labels"]], device=device, dtype=torch.float32)
+                labels_tensor = torch.tensor([tokenized["labels"]], device=device, dtype=torch.bfloat16)
 
                 # ✅ FIX: Create attention mask with the correct integer dtype (torch.long).
                 # Transformer models expect an integer mask, not a float one.
@@ -77,9 +77,9 @@ def evaluate_model(model, tokenizer, dataset, device, max_length, step_separator
                     attention_mask=attention_mask
                 )
 
-                # Ensure consistent float32 dtype throughout
+                # Ensure consistent bfloat16 dtype throughout
                 predictions = outputs.logits.to(torch.float32).cpu().numpy()[0]  # (seq_len,)
-                labels = labels_tensor.cpu().numpy()[0]  # (seq_len,)
+                labels = labels_tensor.to(torch.float32).cpu().numpy()[0]  # (seq_len,)
 
                 # Extract only the non-ignored labels and corresponding predictions
                 valid_mask = labels != -100.0
